Modify app creation to work with azure web.config
