+++ web.config
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <configuration>
   <appSettings>
-    <add key="WSGI_ALT_VIRTUALENV_HANDLER" value="middleware.views.app" />
+    <add key="WSGI_ALT_VIRTUALENV_HANDLER" value="middleware.app.app" />
     <add key="WSGI_ALT_VIRTUALENV_ACTIVATE_THIS"
          value="D:\home\site\wwwroot\env\Scripts\python.exe" />
     <add key="WSGI_HANDLER"
